+++ packages/electrode-react-webapp/lib/jsx/Component.js
@@ -1,0 +1,18 @@
+"use strict";
+
+class Component {
+  constructor(props, context) {
+    this.props = props;
+    this.context = context;
+  }
+
+  isComponent() {
+    return true;
+  }
+
+  render() {
+    return "component";
+  }
+}
+
+module.exports = Component;