+++ packages/electrode-react-webapp/lib/jsx/IndexPage.js
@@ -1,0 +1,11 @@
+"use strict";
+
+const Component = require("./Component");
+
+class IndexPage extends Component {
+  static memoize(props) {
+    return `<!DOCTYPE ${props.DOCTYPE || "html"}>`;
+  }
+}
+
+module.exports = IndexPage;